Added two addons
These are addons I use with my multiboxing setup. I designed it to make it easier for me to get going and to always keep every character in sync.

The BartenderToggle addon is just a simple slash command to show or hide all action bars

The Interfaced addon sets the location of the player and target frames and hides the party frames