Interfaced
Removed the moving of the cast bar. There were quite a few issues with this when the player entered and exited vehicles, and it's easier just to stop moving it. It still looks acceptable. I might be moving it elsewhere on the screen in the future, but not attached to the player frame.